server/src/websocket/trading.ts
import WebSocket, { Server } from 'ws';
import { WebSocketClient, OrderMessage } from '../types/websocket';
import WebSocketBase from './websocket';

class TradingWebSocket extends WebSocketBase {
  protected orders: Map<string, any>;

  constructor(server: Server) {
    super(server);
    this.orders = new Map();
  }

  protected async handleMessage(client: WebSocketClient, data: OrderMessage): Promise<void> {
    try {
      switch (data.type) {
        case 'placeOrder':
          if (data.symbol && data.orderType && data.side && data.quantity) {
            await this.handlePlaceOrder(client, data);
          } else {
            this.sendError(client, 'Invalid place order message: missing required fields');
          }
          break;
        case 'cancelOrder':
          if (data.orderId) {
            await this.handleCancelOrder(client, data.orderId);
          } else {
            this.sendError(client, 'Invalid cancel order message: missing orderId');
          }
          break;
        case 'modifyOrder':
          if (data.orderId) {
            await this.handleModifyOrder(client, data);
          } else {
            this.sendError(client, 'Invalid modify order message: missing orderId');
          }
          break;
        default:
          this.sendError(client, `Unknown message type: ${data.type}`);
      }
    } catch (error) {
      this.sendError(client, error.message);
    }
  }

  protected async handlePlaceOrder(client: WebSocketClient, order: OrderMessage): Promise<void> {
    try {
      // Implement order placement logic here
      const orderId = Math.random().toString(36).substring(7);

      client.ws.send(JSON.stringify({
        type: 'order_placed',
        orderId,
        order
      }));
    } catch (error) {
      this.sendError(client, 'Failed to place order');
    }
  }

  protected async handleCancelOrder(client: WebSocketClient, orderId: string): Promise<void> {
    try {
      // Implement order cancellation logic here
      client.ws.send(JSON.stringify({
        type: 'order_cancelled',
        orderId
      }));
    } catch (error) {
      this.sendError(client, 'Failed to cancel order');
    }
  }

  protected async handleModifyOrder(client: WebSocketClient, order: OrderMessage): Promise<void> {
    try {
      // Implement order modification logic here
      client.ws.send(JSON.stringify({
        type: 'order_modified',
        orderId: order.orderId,
        order
      }));
    } catch (error) {
      this.sendError(client, 'Failed to modify order');
    }
  }

  protected handleDisconnect(client: WebSocketClient): void {
    super.handleDisconnect(client);
    this.clients.delete(client.id);
  }

  protected sendError(client: WebSocketClient, message: string): void {
    super.sendError(client, message);
  }
}

export default TradingWebSocket;
